+++ lib/shoulda/matchers/action_controller.rb
@@ -11,6 +11,8 @@ require 'shoulda/matchers/action_controller/rescue_from_matcher'
 require 'shoulda/matchers/action_controller/callback_matcher'
 require 'shoulda/matchers/action_controller/strong_parameters_matcher'
 require 'shoulda/matchers/action_controller/set_session_or_flash_matcher'
+require 'shoulda/matchers/action_controller/flash_store'
+require 'shoulda/matchers/action_controller/session_store'
 
 module Shoulda
   module Matchers